Update theme configuration & remove unused layouts
Update the website to use the custom Jekyll theme and remove the
unnecessary layouts, includes and/or other view templates, since they
will be inherited from the theme. Also, update the main CSS as well to
inherit the appropriate styles from the custom Jekyll theme.

This has been done in order to separate theme and layout from content.